refactor: position 찾는 기능 제거
