+++ src/test/java/chess/domain/board/BoardTest.java
@@ -2,7 +2,7 @@ package chess.domain.board;
 
 import chess.domain.piece.King;
 import chess.domain.piece.Pawn;
-import chess.domain.piece.notation.Color;
+import chess.domain.piece.Color;
 import chess.domain.position.Position;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
